Ensures that integers are used for decompressing digitized data

+++ commander3/python/commander_tools/tod_tools/commander_tod.py
@@ -373,7 +373,7 @@ class commander_tod:
                     nmax = self.outFile[field].attrs['max']
 
                     bins = np.linspace(nmin, nmax, num = nbins)
-                    dataBuf = bins[dataBuf]
+                    dataBuf = bins[dataBuf.astype('int')]
 
                 elif comp == 'huffman':
                     pid = field.split('/')[1]
